# Avoid LINQ and array copies in `AggregateBindable`