| Method |       Mean |   Error |  StdDev | Ratio |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |-----------:|--------:|--------:|------:|--------:|------:|------:|----------:|
|    Old | 1,262.4 us | 3.17 us | 2.81 us |  1.00 | 54.6875 |     - |     - |    616 KB |
|    New |   883.0 us | 3.83 us | 3.40 us |  0.70 | 15.6250 |     - |     - |    180 KB |

Much of the remaining overhead that is part of the benchmark is coming from locking and `WeakReference.TryGetValue`. Just wanted to get rid of the LINQ as it felt like a bad smell in a path which is actually relatively hot these days.